Game Enhancements and Bug Fixes
Doubled the duration of in-game days.

Introduced four distinct music types corresponding to dawn, day, dusk, and night.

Enriched the game with additional sounds and music tracks.

Enhanced visual appeal with the addition of new textures.

Upgraded all buildings to support up to 15 levels.

Implemented the strokeText function to improve text readability.

Resolved an issue causing night music to play during the day.

Fixed a bug that prevented the saving of building workers.

Renamed ResourceClick to GatheredNumber for clarity.

Ensured functionality of all cards.

Corrected a text color issue in the chest shop.