Architect Operational and BuyButton Enhanced
Architect is now fully operational, enabling placement of objects on the map.

Alteration in the storage method for mapObjects has been implemented.

The cost of chests has been increased, however, this can be offset by reducing the vault level.

Enhancements to the BuyButton now support multiple images and allow for custom positioning of these images within the text.