Enhancements and Bug Fixes
Updated images for architects and craftsmen.

Modified images for architect objects and added corresponding names.

Adjusted resource quantity in the shop for better focus on additional percentages.

Resolved an issue with the text size in the BuyButton.

Expanded the use of BuyButton to more buttons.

Implemented onResize logic for certain elements.

Corrected color issues with the cost of architect objects.

Altered the shadow effect for maxed cards.

Fixed a bug related to card levels.

Updated shop resources to be based on normal and critical gathering.